+++ llarp/exit/endpoint.hpp
@@ -2,7 +2,7 @@
 #define LLARP_EXIT_ENDPOINT_HPP
 
 #include <crypto/types.hpp>
-#include <net/ip.hpp>
+#include <net/ip_packet.hpp>
 #include <path/path.hpp>
 #include <util/time.hpp>
