+++ llarp/exit/session.hpp
@@ -2,7 +2,7 @@
 #define LLARP_EXIT_SESSION_HPP
 
 #include <exit/exit_messages.hpp>
-#include <net/ip.hpp>
+#include <net/ip_packet.hpp>
 #include <path/pathbuilder.hpp>
 #include <routing/transfer_traffic_message.hpp>
